+++ src/main/resources/templates/front/mypage/deliveryAddressList.html
@@ -21,9 +21,9 @@
                 <div class="space-y-4">
                     <div class="space-y-2">
                         <label class="text-sm font-medium leading-none peer-disabled:cursor-not-allowed peer-disabled:opacity-70" for="memberNumber">
-                            회원 번호 <!-- 회원 번호 라벨 -->
+                            회원 번호 <!-- 회원 번호 라벨 데이터베이스 인덱스 번호-->
                         </label>
-                        <input class="flex h-10 w-full rounded-md border border-input bg-background px-3 py-2 text-sm ring-offset-background file:border-0 file:bg-transparent file:text-sm file:font-medium placeholder:text-muted-foreground focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:cursor-not-allowed disabled:opacity-50" id="memberNumber" placeholder="회원 번호를 입력하세요" /> <!-- 회원 번호 입력란 -->
+                        <input class="flex h-10 w-full rounded-md border border-input bg-background px-3 py-2 text-sm ring-offset-background file:border-0 file:bg-transparent file:text-sm file:font-medium placeholder:text-muted-foreground focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:cursor-not-allowed disabled:opacity-50" id="memberNumber" placeholder="회원 번호 데이터베이스 인덱스 번호" /> <!-- 인덱스 번호 -->
                     </div>
                     <div class="space-y-2">
                         <label class="text-sm font-medium leading-none peer-disabled:cursor-not-allowed peer-disabled:opacity-70" for="recipient">
