+++ src/main/resources/templates/front/mypage/lentalList.html
@@ -3,10 +3,8 @@
       xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout"
       layout:decorate="~{front/layouts/main}">
 <main layout:fragment="content">
-페이렌탈리스트
+
+    렌탈리스트
 </main>
+
 </html>
-
-
-
-
